requirements.txt
python==3.7
pytorch==1.9
numpy==1.15.1